DDB Monster Parser refactor, WIP
