+++ src/parser/monster/conditions.js
@@ -1,5 +1,5 @@
 import DICTIONARY from "../../dictionary.js";
-import DDBMonster from "./DDBMonster.js";
+import DDBMonster from "../DDBMonster.js";
 
 DDBMonster.prototype.getAdjustmentsConfig = function getAdjustmentsConfig(type) {
   const damageAdjustments = CONFIG.DDB.damageAdjustments;
