+++ src/parser/monster/senses.js
@@ -1,5 +1,5 @@
 import DICTIONARY from "../../dictionary.js";
-import DDBMonster from "./DDBMonster.js";
+import DDBMonster from "../DDBMonster.js";
 
 DDBMonster.prototype.getTextSenses = function getTextSenses() {
   return this.source.sensesHtml;
