+++ src/parser/monster/size.js
@@ -14,7 +14,7 @@
 
 import DICTIONARY from '../../dictionary.js';
 import logger from '../../logger.js';
-import DDBMonster from "./DDBMonster.js";
+import DDBMonster from "../DDBMonster.js";
 
 DDBMonster.prototype.getSizeFromId = function getSizeFromId(sizeId) {
   const size = CONFIG.DDB.creatureSizes.find((s) => s.id == sizeId).name;
